tests(helpers): separate misc functions
